requirements.txt
pandas>=2.0              
numpy>=1.24              
requests>=2.31           


psycopg2-binary>=2.9     
SQLAlchemy>=1.4.36,<2.0  


#apache-airflow==2.10.5   
python-dotenv>=1.0       


# kafka-python>=2.0
# or: confluent-kafka>=2.0


# pyspark>=3.4
# beautifulsoup4>=4.12
# lxml>=5.2